+++ nengo/tests/test_pytest.py
@@ -111,3 +111,18 @@ def test_pyargs(testdir):
     assert "passed" not in outcomes
     assert outcomes["skipped"] > 350
     assert outcomes["deselected"] > 700
+
+
+@pytest.mark.parametrize("param", (True, False))
+def test_allclose(param, allclose):
+    # if the nengo_test_tolerances in setup.cfg are working properly, then
+    # atol will be set such that all these assertions pass
+
+    if param:
+        assert allclose(0, 0.5, atol=0)
+    else:
+        assert allclose(0, 0.25, atol=0)
+    assert allclose(0, 1.9, atol=0)
+    assert allclose(0, 0.9, atol=0)
+    assert allclose(0, 0.9, atol=0)
+    assert not allclose(0, 1.9)